Ok
I actually have movement and a VERY basic gun script working now. The gun script is really modular so it can be used on basically any kind of weapon except projectile because it is raycast. The script has public floats so that other enemies can still use the same health system and damage system as well.